+++ torch_frame/data/stats.py
@@ -29,6 +29,8 @@ class StatType(Enum):
             return [
                 StatType.COUNT,
             ]
+        elif stype == torch_frame.text_embedded:
+            return []
 
         raise NotImplementedError(f"Invalid semantic type '{stype.value}'")