+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Added `TabNet` ([#35](https://github.com/pyg-team/pytorch-frame/pull/35))
+- Added text embedded stype, mapper and encoder. ([#78](https://github.com/pyg-team/pytorch-frame/pull/78))
 - Added `ExcelFormer` example. ([#46](https://github.com/pyg-team/pytorch-frame/pull/46))
 - Support inductive `DataFrame` to `TensorFrame` transformation ([#75](https://github.com/pyg-team/pytorch-frame/pull/75))
 - Added `CatBoost` baseline and tuned `CatBoost` example. ([#73](https://github.com/pyg-team/pytorch-frame/pull/73))